[refactor] Change to an existing query, Add log, Convert type from Long to Int

+++ src/main/java/mogakco/StudyManagement/repository/MemberScheduleRepository.java
@@ -17,5 +17,5 @@ public interface MemberScheduleRepository extends JpaRepository<MemberSchedule, 
 
     Page<MemberSchedule> findAllBySchedule(Schedule schedule, Pageable pageable);
 
-    long countByMember(Member member);
+    Integer countByMember(Member member);
 }
